requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asyncio==3.4.3
blobfile==3.0.0
boto3==1.38.7
botocore==1.38.7
cachetools==5.5.2
cassandra-driver==3.29.2
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
dnspython==2.7.0
docutils==0.21.2
environs==11.2.1
# faiss-cpu==1.10.0  # Removed - replaced with S3 Vector Search
fastapi==0.115.5
fastapi_cors==0.0.6
filelock==3.16.1
fire==0.7.0
fsspec==2024.12.0
python-multipart==0.0.9
future==1.0.0
geographiclib==2.0
geohash2==1.1
geolib==1.0.7
geomet==0.2.1.post1
email-validator==2.1.0
geopy==2.4.1
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-texttospeech==2.26.0
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.27.0
idna==3.10
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.28
langgraph==0.2.60
langgraph-checkpoint==2.0.9
langgraph-sdk==0.1.48
langsmith==0.2.4
llama_models==0.0.63
llama_stack==0.0.63
llama_stack_client==0.0.63
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numpy==2.2.0
ollama==0.4.4
openai==1.57.4
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pillow==11.0.0
prompt_toolkit==3.0.48
proto-plus==1.26.1
protobuf==6.30.2
pyaml==24.12.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodomex==3.21.0
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
pymongo==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9.1
s3transfer==0.12.0
safetensors==0.4.5
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
tenacity==9.0.0
termcolor==2.5.0
threadpoolctl==3.6.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.47.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
wcwidth==0.2.13
SpeechRecognition==3.10.0
pydub==0.25.1
slowapi==0.1.9
PyJWT==2.8.0
